feat: Get a post with id -> Get every posts

+++ model/post.go
@@ -1,22 +1,22 @@
 package model
 
 type Post struct {
-	ID        int      `json:"id"`
-	UserID    string   `json:"user_id"`
-	Text      string   `json:"text"`
-	ImagePath string   `json:"image_path"`
-	Location  [2]int   `json:"location"` // [x, y] 좌표
-	Category  *string  `json:"category"` // NULL 가능
-	Accuracy  *float64 `json:"accuracy"` // NULL 가능
-	CreatedAt string   `json:"created_at"`
+	ID        int        `json:"id"`
+	UserID    string     `json:"user_id"`
+	Text      string     `json:"text"`
+	ImagePath string     `json:"image_path"`
+	Location  [2]float64 `json:"location"` // [x, y] 좌표
+	Category  *string    `json:"category"` // NULL 가능
+	Accuracy  *float64   `json:"accuracy"` // NULL 가능
+	CreatedAt string     `json:"created_at"`
 }
 
 type FilteredPost struct {
-	ID        int      `json:"id"`
-	UserID    string   `json:"user_id"`
-	Text      string   `json:"text"`
-	ImagePath string   `json:"image_path"`
-	Location  [2]int   `json:"location"` // [x, y] 좌표
-	Category  *string  `json:"category"` // NULL 가능
-	Accuracy  *float64 `json:"accuracy"` // NULL 가능
+	ID        int        `json:"id"`
+	UserID    string     `json:"user_id"`
+	Text      string     `json:"text"`
+	ImagePath string     `json:"image_path"`
+	Location  [2]float64 `json:"location"` // [x, y] 좌표
+	Category  *string    `json:"category"` // NULL 가능
+	Accuracy  *float64   `json:"accuracy"` // NULL 가능
 }
